pyproject.toml
[tool.poetry]
name = "binance"
version = "0.1.0"
description = ""
authors = ["nsamardzic13 <nikola.samardzic1997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
requests = "^2.32.3"
pyyaml = "^6.0.1"
boto3 = "^1.34.151"
google-cloud-bigquery = "^3.25.0"
google-auth = "^2.33.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
